Moves BrowserAction and BrowserActionBadge into navigator folder
Auditors: @bsclifton
